Reimplement how PDF is generated
Thanks to help from one of the DEVONthink developers in response to my
posting on the user forum,

https://discourse.devontechnologies.com/t/difference-in-behavior-between-convert-menu-gear-icon/56820/6?u=mhucka

and another posting in which he provided code for a related problem,

https://discourse.devontechnologies.com/t/any-way-to-batch-convert-files-to-searchable-pdfs/14250/2?u=mhucka

I was able to overcome the problem that my previous code did not
produce the PDF from the version of the web page already loaded in the
window -- it produced it apparently anew, which meant that the result
was produced without any session or login data.  The new code gets the
content of the loaded page and works as expected.