Try to improve behavior on Twitter
In an effort to improve captures of tweets, I added awful hacks like
scrolling to the top *twice*. Sometimes Twitter with in particular,
attempting to go back to the top doesn't, and ends up in some random
location.  Trying to go back to the top a second time actually seems
to help, as stupid as that seems.