Fix getting file path
Things worked when using new from template in Dt, but not when I used
my "Creation document ..." applescript.  For some reason, it needs
"path of first item of _record", which doesn't make sense to me
because I don't think _record is a list, but whatever, I'm tired of
fighting AppleScript all the time.