Add code to set the comment in DEVONthink
Something has changed recently in my environment, and DEVONthink no
longer "notices" when a Finder comment is added externally by
Zowie. I'm at a loss to explain what could have changed (it was not a
software update, and the behavior persisted after rebooting the
computer and restarting DEVONthink). This is a hack, but I added code
to ask the Finder itself for the comment string and tell explicitly
DEVONthink to set the comment after running Zowie.