fix: rewrite using script objects
The problem is that using the Foundation framework results in
something being wonky in the compiled version, and the compiled
version does not work properly in an DEVONthink smart rule (even
though it works from a debugger). Using script objects to wrap the
code that needs the Foundation framework, and avoiding the import of
the Foundation framework for the whole file, makes it work in
DEVONthink smart rules.